+++ CHANGES.md
@@ -66,6 +66,7 @@ Authorization:
 -   Added per-record authorization around the `/records/summary/<recordid>` endpoint matching the `/records/<id>` one.
 -   Added per-record authorization around the `/records/summary` endpoint matching the `/records` one.
 -   Added per-record authorization around the `/records/<recordid>/history` endpoint
+-   Added per-record authorization around the `/records/<recordid>/history/<eventId>` endpoint
 
 Others: